+++ src/YamlCheckCommand.php
@@ -3,6 +3,7 @@
 namespace YamlAlphabeticalChecker;
 
 use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
@@ -11,68 +12,102 @@ use Symfony\Component\Yaml\Exception\ParseException;
 class YamlCheckCommand extends Command
 {
     const
-        OPTION_DIR = 'dir',
-        OPTION_SHOW_DIFF = 'diff';
+        ARGUMENT_DIRS_OR_FILES = 'dirsOrFiles',
+        OPTION_EXCLUDE = 'exclude';
 
     protected function configure()
     {
         $this
             ->setName('yaml-alphabetical-check')
             ->setDescription('Check if yaml files is alphabetically sorted')
-            ->addOption(self::OPTION_DIR, null, InputOption::VALUE_REQUIRED | InputOption::VALUE_IS_ARRAY, 'Directories to check')
-            ->addOption(self::OPTION_SHOW_DIFF, null, InputOption::VALUE_NONE, 'Show difference in yaml file');
+            ->addArgument(self::ARGUMENT_DIRS_OR_FILES, InputArgument::REQUIRED | InputArgument::IS_ARRAY, 'Paths to directories or files to check')
+            ->addOption(self::OPTION_EXCLUDE, null, InputOption::VALUE_REQUIRED | InputOption::VALUE_IS_ARRAY, 'Exclude file mask from check');
     }
 
     /**
      * @param \Symfony\Component\Console\Input\InputInterface $input
      * @param \Symfony\Component\Console\Output\OutputInterface $output
-     * @return int|null
+     * @return int
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $output->writeln('<fg=green>Start checking yaml files.</fg=green>');
-        $output->writeln('');
+        Reporting::startTiming();
+
+        $dirsOrFiles = $input->getArgument(self::ARGUMENT_DIRS_OR_FILES);
+        $excludedFileMasks = $input->getOption(self::OPTION_EXCLUDE);
+        $pathToYamlFiles = YamlFilesPathService::getPathToYamlFiles($dirsOrFiles);
+        $processOutput = new ProcessOutput(count($pathToYamlFiles));
 
         $yamlAlphabeticalChecker = new YamlAlphabeticalChecker();
-        $pathToYamlFiles = YamlFilesPathService::getPathToYamlFiles($input->getOption(self::OPTION_DIR));
-        $isShowDiffOptionEnabled = $input->getOption(self::OPTION_SHOW_DIFF) === true;
-        $errors = [];
+        $results = [];
 
         foreach ($pathToYamlFiles as $pathToYamlFile) {
-            $output->write(sprintf('Checking %s: ', $pathToYamlFile));
+            if ($this->isFileSkipped($pathToYamlFile, $excludedFileMasks)) {
+                $output->write($processOutput->process(ProcessOutput::STATUS_CODE_SKIPP));
+                continue;
+            }
 
             if (!is_readable($pathToYamlFile)) {
-                $output->write(sprintf('<fg=red>File is not readable.</fg=red>'));
-                return 1;
+                $message = 'File is not readable.';
+                $results[] = new Result($pathToYamlFile, $message, Result::RESULT_CODE_GENERAL_ERROR);
+                $output->write($processOutput->process(ProcessOutput::STATUS_CODE_ERROR));
             }
 
             try {
                 $sortCheckResult = $yamlAlphabeticalChecker->isDataSorted($pathToYamlFile);
 
                 if ($sortCheckResult) {
-                    $output->writeln('<fg=green>OK</fg=green>');
+                    $output->write($processOutput->process(ProcessOutput::STATUS_CODE_OK));
                 } else {
-                    if ($isShowDiffOptionEnabled) {
-                        $output->write($yamlAlphabeticalChecker->getDifference($pathToYamlFile));
-                        return 2;
-                    }
-
-                    $errors[] = $pathToYamlFile;
-                    $output->writeln('<fg=red>ERROR</fg=red>');
+                    $diff = $yamlAlphabeticalChecker->getDifference($pathToYamlFile);
+                    $results[] = new Result($pathToYamlFile, $diff, Result::RESULT_CODE_INVALID_SORT);
+                    $output->write($processOutput->process(ProcessOutput::STATUS_CODE_INVALID_SORT));
                 }
             } catch (ParseException $e) {
-                $output->writeln(sprintf('Unable to parse the YAML string: %s', $e->getMessage()));
-                return 4;
+                $message = sprintf('Unable to parse the YAML string: %s', $e->getMessage());
+                $results[] = new Result($pathToYamlFile, $message, Result::RESULT_CODE_GENERAL_ERROR);
+                $output->write($processOutput->process(ProcessOutput::STATUS_CODE_ERROR));
+            }
+        }
+        $output->writeln($processOutput->getLegend());
+
+        return $this->printOutput($output, $results);
+    }
+
+    /**
+     * @param \Symfony\Component\Console\Output\OutputInterface $output
+     * @param \YamlAlphabeticalChecker\Result[] $results
+     * @return int
+     */
+    private function printOutput(OutputInterface $output, array $results)
+    {
+        $resultCode = 0;
+
+        foreach ($results as $result) {
+            $resultCode = $result->getResultCode() > $resultCode ? $result->getResultCode() : $resultCode;
+            $output->writeln(sprintf('FILE: %s', $result->getPathToFile()));
+            $output->writeln('-------------------------------------------------');
+            $output->writeln($result->getMessage() . PHP_EOL);
+        }
+
+        $output->writeln(Reporting::printRunTime());
+
+        return $resultCode;
+    }
+
+    /**
+     * @param string $pathToFile
+     * @param array $excludedFileMasks
+     * @return bool
+     */
+    private function isFileSkipped($pathToFile, array $excludedFileMasks = [])
+    {
+        foreach ($excludedFileMasks as $excludedFileMask) {
+            if (strpos($pathToFile, $excludedFileMask) !== false) {
+                return true;
             }
         }
 
-        $output->writeln('');
-        $output->writeln('<fg=green>End of checking yaml files.</fg=green>');
-
-        if (count($errors) > 0) {
-            return 2;
-        }
-
-        return 0;
+        return false;
     }
 }