+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## [Unreleased]
 
+## [2.0.0] - 2017-11-18
+### Added
+- add new option for exclude files with mask, eg. `--exclude=service`
+
+### Changed
+- option directories changed to required argument
+- now you can send file path and directories for argument to check
+
+### Removed
+- remove `classic` mode and replace him for `diff` mode
+
 ## [1.0.1] - 2017-11-05
 ### Fixed
 - fix return exit code in classic mode
@@ -11,5 +22,6 @@
 - create base command to check yaml sort
 - create `--diff` mode
 
-[Unreleased]: https://github.com/sspooky13/yaml-alphabetical-checker/compare/1.0.1...HEAD
+[Unreleased]: https://github.com/sspooky13/yaml-alphabetical-checker/compare/2.0.0...HEAD
+[2.0.0]: https://github.com/sspooky13/yaml-alphabetical-checker/compare/1.0.1...2.0.0
 [1.0.1]: https://github.com/sspooky13/yaml-alphabetical-checker/compare/1.0.0...1.0.1